Added sockets/DelayedStream and proxy/Delayer pseudo-proxy.
DelayedStream provides a wrapper to delay send and recv operations on a stream. It uses two internal inter-connected TCP sockets in order not to break the ability to put the UnderlyingSocket in a SocketSet.
Delayer is a pseudo-proxy that simply wraps the connections that goes through it in a DelayedStream.

TCPServer::GetLocalPort() and TCPSocket::GetLocalPort() have been added, they are notably used to create the internal inter-connected sockets.

Delayer has been added to proxychain (-d, --delay options).