Got rid of the UnderlyingSocket() methods; added a Waitable interface.
A Waitable interface has been added for anything that can be placed in a SocketSet -- even objects that actually have several sockets.
The UnderlyingSocket() methods that were used to get to the Socket and add it to a SocketSet have now been replaced by a RegisterSockets() method that take a nice SocketSetRegistrar object.